pyproject.toml
[project]
name = "cam-stream"
version = "0.1.0"
description = "basic pi cam streaming webapp"
authors = [
    {name = "Dan-Zed",email = "daniel@more.science"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "smbus2 (>=0.5.0,<0.6.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "flask (>=3.0.0,<4.0.0)"
]

[project.optional-dependencies]
dev = [
    "pytest (>=7.0.0,<8.0.0)",
    "pytest-cov (>=4.0.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
